Per POSIX clock_gettime returns an int

+++ rdtsc_rand.h
@@ -23,8 +23,8 @@ static inline uint64_t splitmix64_hash(uint64_t x) {
 static uint64_t rdtsc_nanos() {
 	struct timespec ts;
 
-	int8_t ok = clock_gettime(CLOCK_MONOTONIC, &ts);  // Uptime
-	//int8_t ok = clock_gettime(CLOCK_REALTIME, &ts); // Since epoch
+	int ok = clock_gettime(CLOCK_MONOTONIC, &ts);  // Uptime
+	//int ok = clock_gettime(CLOCK_REALTIME, &ts); // Since epoch
 
 	if (ok != 0) {
 		return 0; // Return 0 on failure (you can handle this differently)
